+++ Modules/Intent.Modules.AspNetCore/Templates/Startup/StartupTemplateRegistration.cs
@@ -22,24 +22,10 @@ namespace Intent.Modules.AspNetCore.Templates.Startup
     {
         public override string TemplateId => StartupTemplate.TemplateId;
 
-        public override ITemplate CreateTemplateInstance(IOutputTarget project)
+        [IntentManaged(Mode.Fully, Body = Mode.Ignore)]
+        public override ITemplate CreateTemplateInstance(IOutputTarget outputTarget)
         {
-            return new StartupTemplate(project, null);
+            return new StartupTemplate(outputTarget, null);
         }
-
-        //public void DoRegistration(ITemplateInstanceRegistry registry, IApplication application)
-        //{
-        //    var targetProjectIds = new List<string>
-        //    {
-        //        VisualStudioProjectTypeIds.CoreWebApp
-        //    };
-
-        //    var projects = application.Projects.Where(p => targetProjectIds.Contains(p.ProjectType.Id));
-
-        //    foreach (var project in projects)
-        //    {
-        //        registry.Register(TemplateId, project, p => new CoreWebStartupTemplate(project, application.EventDispatcher));
-        //    }
-        //}
     }
 }